Add license, and notice in readme, and in footer
Sadly, we cannot edit all our content files and add license headers
everywhere. But at least we can put the license into the repository and
say that we use it.

I've added a custom section to the readme explaining that not
necessarily all files are subject to LGPL, but only content generated by
and for the website are. This is a bit messy, but I'm not sure how to
better do it. Suggestions?

Of course, we have to take care, if we take for example media files
(images, ...) that are released under a different license which does not
allow redistribution under LGPL-3.0, to annotate that accordingly where
we use the file (e.g. image byline). I don't know if this applies to any
file for now, but I think we can fix that once we notice it.